feat: abi method deploy
