+++ src/features/app-interfaces/components/create/deploy-app.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useMemo, useState } from 'react'
-import { useCreateAppInterfaceStateMachine } from '@/features/app-interfaces/data/create-app-interface'
+import { useCreateAppInterfaceStateMachine } from '@/features/app-interfaces/data'
 import { Button } from '@/features/common/components/button'
 import {
   BuildableTransactionType,
@@ -20,7 +20,7 @@ import { algorandClient } from '@/features/common/data/algo-client'
 import { isArc32AppSpec } from '@/features/common/utils'
 import { asAppCallTransactionParams, asMethodCallParams } from '@/features/transaction-wizard/mappers'
 import { asApplicationAbiMethods } from '@/features/applications/mappers'
-import { Arc32AppSpec, TemplateParamType } from '../data/types'
+import { Arc32AppSpec, TemplateParamType } from '../../data/types'
 import { CreateOnComplete } from '@algorandfoundation/algokit-utils/types/app-factory'
 import { AppClientBareCallParams, AppClientMethodCallParams } from '@algorandfoundation/algokit-utils/types/app-client'
 import { MethodDefinition } from '@/features/applications/models'
@@ -30,8 +30,6 @@ import { base64ToBytes } from '@/utils/base64-to-bytes'
 type Props = {
   machine: ReturnType<typeof useCreateAppInterfaceStateMachine>
 }
-
-// TODO: NC - Rename the WIP components + get structure in the correct place
 
 const getTealTemplateParams = (templateParams: ReturnType<typeof useCreateAppInterfaceStateMachine>[0]['context']['templateParams']) => {
   if (!templateParams) {
@@ -57,7 +55,7 @@ const getTealTemplateParams = (templateParams: ReturnType<typeof useCreateAppInt
   )
 }
 
-export function WIPDeployApp({ machine }: Props) {
+export function DeployApp({ machine }: Props) {
   const [state, send] = machine
   invariant(state.context.appSpec && isArc32AppSpec(state.context.appSpec), 'ARC32 app spec is required')
   const [transaction, setTransaction] = useState<BuildAppCallTransactionResult | BuildMethodCallTransactionResult | undefined>(undefined)
