+++ src/features/app-interfaces/components/labels.ts
@@ -3,5 +3,5 @@ export const methodsLabel = 'Methods'
 export const appIdLabel = 'App ID'
 export const createAppInterfaceLabel = 'Create App Interface'
 export const deleteAppInterfaceLabel = 'Delete App Interface'
-export const deployAppLabel = 'Deploy App'
+export const deployAppLabel = 'Deploy new'
 export const deployButtonLabel = 'Deploy'
