+++ src/features/app-interfaces/data/create.ts
@@ -1,9 +1,76 @@
+import { AppSpecVersion } from '@/features/app-interfaces/data/types'
+import { ApplicationId } from '@/features/applications/data/types.ts'
+import { AppInterfaceEntity, DbConnection, dbConnectionAtom } from '@/features/common/data/indexed-db'
+import { loadable, useAtomCallback } from 'jotai/utils'
+import { useCallback } from 'react'
+import { invariant } from '@/utils/invariant'
+import { createTimestamp } from '@/features/common/data'
+import { getAppInterfaces } from '@/features/app-interfaces/data'
 import { assign, setup } from 'xstate'
 import { Arc32AppSpec, Arc4AppSpec, CreateAppInterfaceContext, TemplateParam } from './types'
-import { ApplicationId } from '@/features/applications/data/types'
 import { useMemo } from 'react'
 import { atomWithMachine } from 'jotai-xstate'
-import { useAtom } from 'jotai'
+import { atom, useAtom, useAtomValue } from 'jotai'
+
+export type AppSpecDetails = {
+  applicationId: ApplicationId
+  name: string
+} & AppSpecVersion
+
+export const writeAppInterface = async (dbConnection: DbConnection, appInterface: AppInterfaceEntity) => {
+  await dbConnection.put('app-interfaces', appInterface)
+}
+
+export const useCreateAppInterface = () => {
+  return useAtomCallback(
+    useCallback(async (get, _, appSpecDetails: AppSpecDetails) => {
+      const { applicationId, name, ...appSpecVersion } = appSpecDetails
+      invariant(
+        appSpecVersion.roundFirstValid === undefined || appSpecVersion.roundFirstValid >= 0,
+        'roundFirstValid must be greater than or equal to 0'
+      )
+      invariant(
+        appSpecVersion.roundLastValid === undefined || appSpecVersion.roundLastValid >= 0,
+        'roundLastValid must be greater than or equal to 0'
+      )
+      if (appSpecVersion.roundFirstValid !== undefined && appSpecVersion.roundLastValid !== undefined) {
+        invariant(appSpecVersion.roundLastValid > appSpecVersion.roundFirstValid, 'roundFirstValid must be greater than roundLastValid')
+      }
+
+      const dbConnection = await get(dbConnectionAtom)
+      const existingAppInterfaces = await getAppInterfaces(dbConnection)
+      invariant(
+        existingAppInterfaces.find((e) => e.applicationId === applicationId) === undefined,
+        `Application ID '${applicationId}' is already associated with another app interface`
+      )
+      invariant(
+        existingAppInterfaces.find((e) => e.name.toLowerCase() === name.toLowerCase()) === undefined,
+        `App interface '${name}' already exists, please choose a different name`
+      )
+
+      await writeAppInterface(dbConnection, {
+        applicationId: applicationId,
+        name: name,
+        appSpecVersions: [appSpecDetails],
+        lastModified: createTimestamp(),
+      })
+    }, [])
+  )
+}
+
+const useAppInterfacesAtom = () => {
+  return useMemo(() => {
+    return atom(async (get) => {
+      const dbConnection = await get(dbConnectionAtom)
+      return await getAppInterfaces(dbConnection)
+    })
+  }, [])
+}
+
+export const useLoadableAppInterfacesAtom = () => {
+  const appInterfacesAtom = useAppInterfacesAtom()
+  return useAtomValue(loadable(appInterfacesAtom))
+}
 
 const createMachine = () =>
   setup({
